Update dependencies and simplify pixel index handling in `ViewPort`
- Upgrade Android Gradle Plugin (AGP) from `8.11.2` to `8.13.2`.
- Upgrade Compose Multiplatform from `1.9.3` to `1.10.0`.
- Simplify `pixelIndex` assignment in `ViewPort.draw` by removing the redundant `toInt()` call on `cps.pixels`.

+++ composeApp/src/commonMain/kotlin/pl/pelotasplus/eyeofbeholder/data/model/ViewPort.kt
@@ -421,7 +421,7 @@ class ViewPort {
                 val cpsIndex = j * cps.width + i
                 if (cpsIndex < 0 || cpsIndex >= cps.pixels.size) continue
 
-                val pixelIndex = cps.pixels[cpsIndex].toInt()
+                val pixelIndex = cps.pixels[cpsIndex]
 
                 // Skip transparent pixels (index 0)
                 if (pixelIndex == 0) {
